test function delete file

+++ TimeUpBackend/src/main/java/sit/int221/TimeUpBackend/service/EventService.java
@@ -285,14 +285,14 @@ public class EventService{
                  if(user.getRoleUser().equals("admin")){
                      System.out.println(event.getFileName());
                      System.out.println(event.getBookingEmail());
-                     return conditionEditEvent(multipartFile, id, sizeByte, event);
+                     return conditionEditEvent(editEventPutDTO , multipartFile, id, sizeByte, event);
                  }
                  else if (user.getRoleUser().equals("student") && event.getBookingEmail().equals(getCurrentAuthentication.getUsername())){
                      if(!(event.getBookingEmail().equals(user.getEmailUser()))){
                          throw  new ResponseStatusException(HttpStatus.FORBIDDEN , "email is not the same as student's email");
                      }
                      else{
-                         return conditionEditEvent(multipartFile, id, sizeByte, event);
+                         return conditionEditEvent(editEventPutDTO , multipartFile, id, sizeByte, event);
                      }
                  }
         }
@@ -302,7 +302,7 @@ public class EventService{
         throw  new ResponseStatusException(HttpStatus.FORBIDDEN , "This email permission denied");
     }
 
-    private ResponseEntity conditionEditEvent(MultipartFile multipartFile, Integer id, int sizeByte, Event event) throws IOException {
+    private ResponseEntity conditionEditEvent(EventPutDto editEventPutDTO , MultipartFile multipartFile, Integer id, int sizeByte, Event event) throws IOException {
         if(multipartFile != null && event.getFileName() == null){
             System.out.println(1);
             storageService.save(multipartFile , id);
@@ -314,10 +314,10 @@ public class EventService{
                 System.out.println(2);
                 eventRepository.saveAndFlush(event);
             }
-            else {
+            else if (multipartFile == null && event.getFileName().equals(editEventPutDTO.getFileName())) {
                 System.out.println(3);
                 storageService.deleteById(id);
-                storageService.save(multipartFile , id);
+                storageService.save(null , id);
                 event.setFileSize(sizeByte);
                 eventRepository.saveAndFlush(event);
             }
